modularising operations
