+++ blog/operations/__init__.py
@@ -1,0 +1,6 @@
+# operations/__init__.py
+# Makes the 'operations' folder a Python package.
+
+from .create_post import create_post
+from .view_posts import view_posts
+from .delete_post import delete_post
